Development

+++ lib/repositories/user_repository.dart
@@ -4,8 +4,12 @@ import 'package:anotador/utils/app_data.dart';
 import 'package:sqflite/sqflite.dart';
 
 class UserRepository {
-  Future<User> insertUser(User user) async {
+  Future<User> insertUser(User user, {bool isOwner = false}) async {
     final db = await AppData.database;
+
+    if (isOwner) {
+      user.id = 0;
+    }
 
     int userId = await db.insert(
       Tables.user,
@@ -30,10 +34,13 @@ class UserRepository {
     });
   }
 
-  Future<void> delete(int userId) async {
+  Future<int> delete(int userId) async {
+    if (userId == 0) {
+      return 0; // Owner user cannot be deleted
+    }
     final db = await AppData.database;
 
-    await db.delete(Tables.user,
+    return await db.delete(Tables.user,
         where: '${Tables.user}_id = ?', whereArgs: [userId]);
   }
 }
